+++ doc/source/conf.py
@@ -311,7 +311,7 @@ texinfo_documents = [
 intersphinx_mapping = {'http://docs.python.org/': None,
                        'http://yt-project.org/doc/': None,
                        'http://docs.scipy.org/doc/numpy/': None,
-                       }
+                       'http://hea-www.cfa.harvard.edu/~jzuhone/soxs/': None}
 
 def remove_module_docstring(app, what, name, obj, options, lines):
     if what == "module" and name in ["photon_list","event_list"]: