+++ pyxsim/photon_list.py
@@ -546,10 +546,10 @@ class PhotonList(object):
             profiles and/or spectra. Should probably only be used for visualization purposes.
             Supply a float here to smooth with a standard deviation with this fraction 
             of the cell or particle size. Default: None
-        prng : integer, :class:`~numpy.random.RandomState` object, or :mod:`~numpy.random`, optional
+        prng : integer or :class:`~numpy.random.RandomState` object 
             A pseudo-random number generator. Typically will only be specified
             if you have a reason to generate the same set of random numbers, such as for a
-            test. Default is the :mod:`numpy.random` module.
+            test. Default is to use the :mod:`numpy.random` module.
 
         Examples
         --------